requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
streamlit==1.28.0
pandas==2.1.1

